+++ app/models/surveys/answer_session.rb
@@ -36,10 +36,20 @@ class AnswerSession < ActiveRecord::Base
     end
   end
 
+  def process_answer_from_api(question, response)
+    answer = atomic_first_or_create_answer(question)
+    if answer.locked?
+      nil
+    else
+      answer.update_response_value_from_api!(response)
+      answer
+    end
+  end
+
   def atomic_first_or_create_answer(question)
     begin
       self.answers.where(question_id: question.id).first_or_create!
-    rescue ActiveRecord::RecordNotUnique
+    rescue ActiveRecord::RecordNotUnique, ActiveRecord::RecordInvalid
       retry
     end
   end